+++ x/perp/client/cli/cli_test.go
@@ -211,7 +211,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.EqualValues(sdk.NewDec(60_000_000_000), reserveAssets.QuoteAssetReserve)
 
 	s.T().Log("A. check trader has no existing positions")
-	_, err = testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	_, err = testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 	s.Error(err, "no position found")
 
 	s.T().Log("B. open position")
@@ -235,7 +235,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.EqualValues(sdk.NewDec(60_001_000_000), reserveAssets.QuoteAssetReserve)
 
 	s.T().Log("B. check vpool balances")
-	queryResp, err := testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	queryResp, err := testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 	s.T().Logf("query response: %+v", queryResp)
 	s.NoError(err)
 	s.EqualValues(user.String(), queryResp.Position.TraderAddress)
@@ -262,7 +262,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.NoError(err)
 
 	s.T().Log("C. check trader position")
-	queryResp, err = testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	queryResp, err = testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 	s.T().Logf("query response: %+v", queryResp)
 	s.NoError(err)
 	s.EqualValues(user.String(), queryResp.Position.TraderAddress)
@@ -296,7 +296,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.EqualValues(sdk.NewDec(60_002_999_900), reserveAssets.QuoteAssetReserve)
 
 	s.T().Log("D. Check trader position")
-	queryResp, err = testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	queryResp, err = testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 	s.T().Logf("query response: %+v", queryResp)
 	s.NoError(err)
 	s.EqualValues(user.String(), queryResp.Position.TraderAddress)
@@ -323,7 +323,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.NotContains(res.String(), "fail")
 
 	s.T().Log("E. Check trader position")
-	queryResp, err = testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	queryResp, err = testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 	s.T().Logf("query response: %+v", queryResp)
 	s.NoError(err)
 	s.EqualValues(user.String(), queryResp.Position.TraderAddress)
@@ -346,7 +346,7 @@ func (s *IntegrationTestSuite) TestOpenPositionsAndCloseCmd() {
 	s.NoError(err)
 
 	s.T().Log("F. check trader position")
-	queryResp, err = testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
+	queryResp, err = testutilcli.QueryPosition(val.ClientCtx, common.Pair_BTC_NUSD, user)
 
 	s.T().Logf("query response: %+v", queryResp)
 	s.Error(err)
@@ -361,7 +361,7 @@ func (s *IntegrationTestSuite) TestPositionEmptyAndClose() {
 	user := s.users[0]
 
 	// verify trader has no position (empty)
-	_, err := testutilcli.QueryTraderPosition(val.ClientCtx, common.Pair_ETH_NUSD, user)
+	_, err := testutilcli.QueryPosition(val.ClientCtx, common.Pair_ETH_NUSD, user)
 	s.Error(err, "no position found")
 
 	// close position should produce error
@@ -372,22 +372,6 @@ func (s *IntegrationTestSuite) TestPositionEmptyAndClose() {
 	}
 	out, _ := sdktestutilcli.ExecTestCLICmd(val.ClientCtx, cli.ClosePositionCmd(), append(args, commonArgs...))
 	s.Contains(out.String(), collections.ErrNotFound.Error())
-}
-
-func (s *IntegrationTestSuite) TestGetPrices() {
-	val := s.network.Validators[0]
-
-	s.T().Log("check vpool balances")
-	reserveAssets, err := testutilcli.QueryVpoolReserveAssets(val.ClientCtx, common.Pair_ETH_NUSD)
-	s.NoError(err)
-	s.EqualValues(sdk.MustNewDecFromStr("10000000"), reserveAssets.BaseAssetReserve)
-	s.EqualValues(sdk.MustNewDecFromStr("60000000000"), reserveAssets.QuoteAssetReserve)
-
-	s.T().Log("check prices")
-	priceInfo, err := testutilcli.QueryBaseAssetPrice(val.ClientCtx, common.Pair_ETH_NUSD, "add", "100")
-	s.T().Logf("priceInfo: %+v", priceInfo)
-	s.EqualValues(sdk.MustNewDecFromStr("599994.000059999400006000"), priceInfo.PriceInQuoteDenom)
-	s.NoError(err)
 }
 
 func (s *IntegrationTestSuite) TestQueryCumulativePremiumFractions() {
@@ -433,6 +417,84 @@ func (s *IntegrationTestSuite) TestRemoveMargin() {
 	}
 
 	s.Contains(out.String(), perptypes.ErrFailedRemoveMarginCanCauseBadDebt.Error())
+}
+
+func (s *IntegrationTestSuite) TestAddMargin() {
+	val := s.network.Validators[0]
+	pair := common.Pair_ETH_NUSD
+
+	// Open a new position
+	s.T().Log("opening a position with user 2....")
+	args := []string{
+		"--from",
+		s.users[2].String(),
+		"buy",
+		pair.String(),
+		"10",    // Leverage
+		"10000", // Quote asset amount
+		"0.0000001",
+	}
+
+	_, err := sdktestutilcli.ExecTestCLICmd(val.ClientCtx, cli.OpenPositionCmd(), append(args, commonArgs...))
+	if err != nil {
+		s.T().Logf("user2 open position err: %+v", err)
+	}
+	s.Require().NoError(err)
+
+	testCases := []struct {
+		name           string
+		args           []string
+		expectedCode   uint32
+		expectedMargin sdk.Dec
+	}{
+		{
+			name: "PASS: add margin to correct position",
+			args: []string{
+				"--from",
+				s.users[2].String(),
+				pair.String(),
+				fmt.Sprintf("%s%s", "10000", pair.Token1),
+			},
+			expectedCode:   0,
+			expectedMargin: sdk.NewDec(20_000),
+		},
+		{
+			name: "FAIL: position not found",
+			args: []string{
+				"--from",
+				s.users[2].String(),
+				common.Pair_BTC_NUSD.String(),
+				fmt.Sprintf("%s%s", "10000", pair.Token1),
+			},
+			expectedCode: 1,
+		},
+	}
+
+	for _, tc := range testCases {
+		s.T().Run(tc.name, func(t *testing.T) {
+			s.T().Log("adding margin on user 2....")
+			out, err := sdktestutilcli.ExecTestCLICmd(val.ClientCtx, cli.AddMarginCmd(), append(tc.args, commonArgs...))
+			if err != nil {
+				s.T().Logf("user 2 add margin err: %+v", err)
+			}
+			s.Require().NoError(err)
+
+			var tx sdk.TxResponse
+			val.ClientCtx.Codec.MustUnmarshalJSON(out.Bytes(), &tx)
+
+			if tc.expectedCode != 0 {
+				s.EqualValues(tc.expectedCode, tx.Code)
+			} else {
+				s.EqualValues(tc.expectedCode, 0)
+
+				// query trader position
+				queryResp, err := testutilcli.QueryPosition(val.ClientCtx, pair, s.users[2])
+				s.NoError(err)
+
+				s.EqualValues(tc.expectedMargin, queryResp.Position.Margin)
+			}
+		})
+	}
 }
 
 func (s *IntegrationTestSuite) TestLiquidate() {