+++ x/perp/types/expected_keepers.go
@@ -103,6 +103,8 @@ type VpoolKeeper interface {
 		lookbackInterval time.Duration,
 	) (quoteAssetAmount sdk.Dec, err error)
 
+	GetAllPools(ctx sdk.Context) []vpooltypes.VPool
+
 	IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) bool
 	GetMaintenanceMarginRatio(ctx sdk.Context, pair common.AssetPair) sdk.Dec
 	GetMaxLeverage(ctx sdk.Context, pair common.AssetPair) sdk.Dec